refactor: move root_directory method to test_utils
