+++ src/queens/utils/remote_operations.py
@@ -29,9 +29,10 @@ import cloudpickle
 from fabric import Connection
 from invoke.exceptions import UnexpectedExit
 
-from queens.utils.path import PATH_TO_ROOT, is_empty
+from queens.utils.path import is_empty
 from queens.utils.rsync import assemble_rsync_command
 from queens.utils.run_subprocess import start_subprocess
+from test_utils.path import PATH_TO_ROOT
 
 _logger = logging.getLogger(__name__)
